CI: Add a couple of periodic jobs using ovn-fake-multinode.
This patch adds a couple of jobs using ovn-fake-multinode.
It first builds 2 ovn-fake-multinode container images
  - one with OVN 22.03
  - one with present main.

The first job deploys ovn-fake-multinode with the main
OVN and runs simple tests provided by ovn-fake-multinode [1].

The second job deploys ovn-fake-multinode setup with the
central image using OVN 22.03 and chassis image using main OVN.
This job tests the scenario
  - ovn-northd and OVN dbs are running the most recent LTS.
  - ovn-controller is running the latest commit from the branch.

The workflow is right now scheduled to trigger on midnight everyday.
Once we cache the built image or reduce the overall run time of
this workflow we can enable for every push.

[1] - https://github.com/ovn-org/ovn-fake-multinode/blob/main/.ci/test_basic.sh

Reviewed-by: Simon Horman <simon.horman@corigine.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>