tests: Fixed system-test "load balancing affinity sessions"
This tests (and potentially others) might fail on some systems
as detecting unexpected flows in conntrack, related to host.
This was due to the fact FORMAT_CT grep for the expected ip address
but understands the "." in the ip address as "any character".
So, for instance FORMAT_CT("192.168.1.2") is true for 192.168.122.1.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>