ovn-trace: need to read latest data when using daemon mode
When logical flow changed in SB, the result that ovn-trace evaluate
using daemon mode is previous. Because the data from db just is read
once when ovn-trace startup. This situation is not adapted for daemon
mode. So we need to keep latest data from SB when ovn-trace use
daemon mode.

In addition, print the unixctl server path used to be recorded refer
from ovn-nbctl and ovn-sbctl.

Signed-off-by: Jun Gu <jun.gu@easystack.cn>