build-aux/sodepends.py: Fix broken build when manpage changes.
Commit [0] updated ovn-detrace.1.in, which triggered an existing problem
of the build system. After this change, for any existed build
environments that have created a separate build dir (instead of directly
build under the root dir of the source code), the build is broken, and
complains:

utilities/ovn-detrace.1.in not found in: . srcdir OVS_MANDIR .
make: *** [Makefile:3482: ../manpages.mk] Error 1

The root cause is that an earlier commit b6a75011 had a bug and the
sodepends script didn't read the new -I input properly.

[0] 25b4d763bd4b ("ovn-detrace: Support connecting to NB and SB raft followers")

Fixes: b6a750112324 ("manpages.mk: fix dependencies path")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>