ovn-nbctl: Add tier ACL options.
This modifies the acl-add and acl-del commands so that an ACL
tier can be specified when adding or deleting ACLs.

For acl-add, if the tier is specified, then the ACL created by the
command will have that tier set.

For acl-del, if the tier is specified, then the tier will be one of the
criteria used when deciding which ACLs to delete. Because the tier is
not any more or less specific than the other criteria used for deleting
ACLs, a bitmap approach is used to determine the final set of ACLs that
should be deleted.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>
Reviewed-by: Ales Musil <amusil@redhat.com>