northd: Expose real inputs through function arguments.
The I-P engine node functions get inputs from the node's input data, and
then call the real implementation functions (implemented in northd.c)
with a xxx_input wrapper variable, and then in the called functions, it
passes the variable further down to other functions, regardless of what
fields are really used by those functions. This is fine when there is no
real I-P implementation (i.e. we are always recompute), but to add
incremental processing, we need to figure out which functions depend
on which part of the input data, so that we can keep track of the logic
that is needed to handle a specific change in the input data. Ideally,
for incremental processing to work, we should be able to draw a
dependency graph based on function prototypes.

To fulfill this purpose, this patch refactors related functions to
expose real inputs through function arguments. It made some of the
function prototypes lengthy (in fact, not too much, because it turned
out most of the functions only require very few members of the wrapper
input data structure), but the dependencies are much more clear.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>