northd: Remove redundant checks.
Commits 53febfbc3776 ("northd: Split switch and router datapaths.") and
b2f09ac55041 ("northd: Split switch ports and router ports.") made it
such that for most router/switch specific functions we never process
records that are not applicable to the function type.  Remove the
redundant checks and replace them with asserts as it's a bug to call the
functions with the wrong types of ports/datapaths.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>