system-tests: Reduce flakiness of netcat UDP clients
When we send a packet through UDP without other side
having open port, it usually replies with ICMP message.
netcat upon getting that error fails with:
"Ncat: Connection refused.". However, there is a race
because this message might arrive after the netcat client
is already done, so it will exit without any error.

To prevent that error ignore the return code, stdout and stderr
for those calls. It is fine to ignore because all those changed
instances are checking packet count after the calls so if the
netcat fails for other reasons it will be detected by the packet
count check.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>