ci: ovn-kubernetes: Pin ovn-k8s version.
Recently, ovn-k8s started using the new OVN Northbound database table
Chassis_Template_Var that was added in v22.12.0.  Unfortunately,
ovn-k8s (libovsdb) became incompatible with OVN 22.09 as it doesn't
support connecting to servers that don't contain this table in their
schema.

It's still important to check that we don't introduce regressions on
stable branches so, instead of completely removing the ovn-k8s CI jobs,
we now pin ovn-k8s to a version that's known to work with OVN 22.09.

NOTE: This is not really a permanent solution, it's just a way to
unblock and still run some of the ovn-k8s CI on branches older than
22.09.  The ideal solution would be to ensure that ovn-k8s can work with
older OVN versions (maybe it's enough to ensure compatibility with LTS).
Until then we will have to keep pinning ovn-kubernetes on stable
branches every time a new table is used by the CMS.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Reviewed-by: Simon Horman <simon.horman@corigine.com>