northd.c: Avoid sending ICMP time exceeded for multicast packets.
In RFC1812 section 5.3.1, it is mentioned that:

   If the TTL is reduced to zero (or less), the packet MUST be
   discarded, and if the destination is not a multicast address the
   router MUST send an ICMP Time Exceeded message ...

So if the destionation is a multicast address the route shouldn't send
ICMP Time Exceeded, but the current OVN implementation didn't check
multicast and tries to send ICMP regardless. This patch fixes it.

The reason behind is that TTL has special meanings for multicast. For
example, TTL = 1 means restricted to the same subnet, not forwarded by
the router. So it is very common to see multicast packets with ttl = 1,
and generating ICMP for such packets is harmful from both slowpath
performance and functionality point of view.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>