tests: check arguments count of OVS_WAIT_UNTIL
The macro has been used erroneously with the expectation
to take the third parameter as the expected output.
Fail if more than 2 arguments are used.
Also fail if the second argument is an integer (second arg expected to
be a command to execute in case of failure).
Same checks are used in ovs.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>