mac-learn: Make the mac_binding struct more flexible
In order to reuse the mac_binding struct make it more flexible
by abstracting the commit in generic timestamp. At the same time
add struct that holds map of mac bindings which allows to specify
the max size during initialization.

This server as preparation for the packet buffering simplification
and race condition fix.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>