northd: Update obsolete datapath groups instead of re-creating.
Every time we add a new switch or router it may result in addition or
removal of a few datapath in datapath groups.  Every time that happens,
northd will remove obsolete groups, create new groups that differ by a
few datapaths, and update all affected Sb LBs and logical flows with
references to these new groups.  This generates large database
transactions that touch a lot of records in the Southbound DB making
Sb DB server generate large updates for all connected clients.

All in all, this behavior leads to high CPU usage in all OVN components
on, for example, attempt to add or remove the node to/from a scaled
OVN-Kubernetes cluster.  It also prevents implementing datapath group
incremental processing in ovn-controller, because groups are never
updated.

Update existing obsolete groups already attached to the logical flow
or Sb load balancer instead of re-creating to avoid these issues.

The logic of datapath group updates was different between LBs and LFs.
And the way it was implemented didn't allow efficient re-use of
existing obsolete groups.  So, it was slightly re-worked, unified
between Sb LBs and logical flows and moved into a separate common
function.

The logic is following:

 1. Iterate over Sb records that reference DP groups and check these
    groups, find existing or create new groups for records that will
    stay in the DB.  If the currently attached group is different from
    the desired one - modify, if it's not already used for something
    else.

 2. Iterate over all the new records and find existing or create new
    groups for them.

This logic allows to avoid group duplication within a resource type
and re-use existing groups as long as possible.

Tests added to check datapath groups for logical flows.  Test for
LB sync updated to make sure that groups are not re-created.

In a test where one new switch with a load balancer group is added
into the cluster at the end of 500-node density-heavy ovn-heater test,
the change allows to reduce the database update size from 11 MB
down to just 90 KB.  The total time needed to complete this nbctl
operation with --wait=hv goes down from 15-16 seconds to 4 seconds.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Reviewed-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>