utilities: increase OVSDB inactivity probe interval for ovn-*ctl
For large OVN_Southbound (or other) databases the default interval of 5000 ms
could be not sufficient to run.  This patch adds configuration of OVSDB
inactivity probes for ovn-*ctl utilities.

Initially, on the utility start the hardcoded value of 120000 ms is set.
For daemon-mode it is possible to configure intervals in OVN Northbound and
OVN Southbound databases for ovn-nbctl and ovn-sbctl utilities respectively.

Use
OVN_Northbound.NB_Global.options.nbctl_probe_interval for ovn-nbctl and
OVN_Southbound.SB_Global.options.sbctl_probe_interval for ovn-sbctl
utilities.

If no DB configuration option was provided, the initial (120000 ms) interval
is left.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Acked-by: Ales Musil <amusil@redhat.com>